Navbar

+++ src/Utility/Fetch.js
@@ -20,9 +20,10 @@ export async function UserLogin(Account) {
 
 export async function UserLogOut() {
     const url = "/logout"
-    // const {state} = useContext(context)
     let response
     try{
+        const cid = localStorage.getItem('authUser.cid')
+        if(!cid){return}
         response = await fetch(url, { credentials: 'include' })
         console.log(response)
         localStorage.clear()
